+++ CHANGELOG.md
@@ -6,4 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+- Initial release.
+- Run `Checkov` on a file.
+- Display diagnostics errors.
+- Install and update `Checkov`.